+++ autosub/__init__.py
@@ -12,10 +12,10 @@ from __future__ import absolute_import, print_function, unicode_literals
 import pysubs2
 
 # Any changes to the path and your own modules
-from autosub import core
 from autosub import ffmpeg_utils
 from autosub import cmdline_utils
 from autosub import options
+from autosub import exceptions
 
 
 def main():  # pylint: disable=too-many-branches, too-many-statements, too-many-locals
@@ -26,14 +26,15 @@ def main():  # pylint: disable=too-many-branches, too-many-statements, too-many-
     args = options.get_cmd_args()
 
     try:
-        ffmpeg_cmd = ffmpeg_utils.check_cmd("ffmpeg")
+        ffmpeg_cmd = ffmpeg_utils.get_cmd("ffmpeg")
         if not ffmpeg_cmd:
-            raise core.PrintAndStopException(
+            raise exceptions.AutosubException(
                 "Error: Dependency ffmpeg on this machine."
             )
 
+        ffmpeg_cmd = ffmpeg_cmd + ' '
         if cmdline_utils.list_args(args):
-            raise core.PrintAndStopException("\nAll works done.")
+            raise exceptions.AutosubException("\nAll works done.")
 
         if not args.yes:
             try:
@@ -49,42 +50,53 @@ def main():  # pylint: disable=too-many-branches, too-many-statements, too-many-
         if validate_result == 0:
             if args.audio_process:
                 if args.audio_process == 'y':
-                    args.input = ffmpeg_utils.audio_pre_prcs(
+                    prcs_file = ffmpeg_utils.audio_pre_prcs(
                         filename=args.input,
                         is_keep=args.keep,
                         cmds=args.audio_process_cmd,
                         input_m=input_m,
                         ffmpeg_cmd=ffmpeg_cmd
                     )
-                    print("Audio pre-processing complete.")
-                    is_flac = True
+                    if not prcs_file:
+                        no_audio_prcs = False
+                    else:
+                        args.input = prcs_file
+                        print("Audio pre-processing complete.")
+                        no_audio_prcs = True
                 elif args.audio_process == 'o':
                     args.keep = True
-                    args.input = ffmpeg_utils.audio_pre_prcs(
+                    prcs_file = ffmpeg_utils.audio_pre_prcs(
                         filename=args.input,
                         is_keep=args.keep,
                         cmds=args.audio_process_cmd,
                         input_m=input_m,
                         ffmpeg_cmd=ffmpeg_cmd
                     )
-                    raise core.PrintAndStopException(
-                        "Audio pre-processing complete.\nAll works done."
-                    )
+                    if not prcs_file:
+                        raise exceptions.AutosubException(
+                            "No works done."
+                        )
+                    else:
+                        args.input = prcs_file
+                        raise exceptions.AutosubException(
+                            "Audio pre-processing complete.\nAll works done."
+                        )
                 elif args.audio_process == 'n':
-                    is_flac = True
+                    no_audio_prcs = True
                 else:
-                    is_flac = False
+                    no_audio_prcs = False
             else:
-                is_flac = False
+                no_audio_prcs = False
 
             cmdline_utils.validate_aovp_args(args)
-            cmdline_utils.fix_args(args)
+            cmdline_utils.fix_args(args,
+                                   ffmpeg_cmd=ffmpeg_cmd)
             fps = cmdline_utils.get_fps(args=args, input_m=input_m)
             cmdline_utils.audio_or_video_prcs(args,
                                               fps=fps,
                                               input_m=input_m,
                                               styles_list=styles_list,
-                                              is_flac=is_flac)
+                                              no_audio_prcs=no_audio_prcs)
 
         elif validate_result == 1:
             cmdline_utils.validate_sp_args(args)
@@ -100,7 +112,7 @@ def main():  # pylint: disable=too-many-branches, too-many-statements, too-many-
     except pysubs2.exceptions.Pysubs2Error:
         print("\nError: pysubs2.exceptions. Check your file format.")
         return 1
-    except core.PrintAndStopException as err_msg:
+    except exceptions.AutosubException as err_msg:
         print(err_msg)
         return 0